ADD tests for model serilizer and view
